sptests: Address -Wsign-compare warnings
Most of these warnings were between int or ssize_t and size_t.
In particular, various POSIX system calls like read() and write()
return ssize_t and comparing that value with the sizeof a buffer
is a common source of these warnings. Another common source is
using an int as the iterator in a for loop with the limit being
a size_t.

With the type change, some printf() specifiers needed to change also.

+++ testsuites/sptests/sp71/init.c
@@ -50,7 +50,7 @@ rtems_task Init(
 
   rtems_status_code status;
   rtems_id          id;
-  long              created;
+  size_t            created;
 
   TEST_BEGIN();
 
@@ -70,7 +70,7 @@ rtems_task Init(
   } while (1);
 
   printf(
-    "%ld ports created using %d per allocation -- need %ld\n",
+    "%zu ports created using %d per allocation -- need %ld\n",
     created,
     PER_ALLOCATION,
     (long) OBJECTS_ID_FINAL_INDEX
